+++ sopa/cli/app.py
@@ -37,12 +37,6 @@ app.add_typer(
 
 
 @app.command()
-def read(data_path: str = typer.Argument(), technology: str = typer.Option()):
-    """Deprecated and will be removed in sopa==2.1.0. Use `sopa convert` instead."""
-    raise NameError("`sopa read` is deprecated. Use `sopa convert` instead.")
-
-
-@app.command()
 def convert(
     data_path: str = typer.Argument(
         help="Path to one data sample (most of the time, this corresponds to a directory with images files and eventually a transcript file)"
@@ -175,29 +169,22 @@ def aggregate(
         None,
         help="Optional image key of the SpatialData object. By default, considers the only one image. It can be useful if another image is added later on",
     ),
+    gene_column: str = typer.Option(
+        None, help="Optional column of the transcript dataframe to be used as gene names. Inferred by default."
+    ),
     method_name: str = typer.Option(
         None, help="If segmentation was performed with a generic method, this is the name of the method used."
     ),
-    average_intensities: bool = typer.Option(False, help="[Deprecated] Use `aggregate_channels` instead."),
-    gene_column: str = typer.Option(None, help="[Deprecated] Use `aggregate_genes` instead."),
 ):
     """Create an `anndata` table containing the transcript count and/or the channel intensities per cell"""
     import sopa
     from sopa.io.standardize import read_zarr_standardized
 
-    sdata = read_zarr_standardized(sdata_path)
-
-    if gene_column is not None:
-        log.warning(
-            "The `gene_column` argument is deprecated and will be removed in sopa==2.1.0. Use `aggregate_genes` instead."
-        )
+    if gene_column:
+        assert aggregate_genes is not False
         aggregate_genes = True
 
-    if average_intensities:
-        log.warning(
-            "The `average_intensities` argument is deprecated and will be removed in sopa==2.1.0. Use `aggregate_channels` instead."
-        )
-        aggregate_channels = True
+    sdata = read_zarr_standardized(sdata_path)
 
     sopa.aggregate(
         sdata,
@@ -205,6 +192,7 @@ def aggregate(
         aggregate_channels=aggregate_channels,
         image_key=image_key,
         shapes_key=method_name,
+        gene_column=gene_column,
         min_transcripts=min_transcripts,
         expand_radius_ratio=expand_radius_ratio,
         min_intensity_ratio=min_intensity_ratio,