+++ CHANGELOG.md
@@ -1,10 +1,21 @@
-## [2.0.8] - xxxx-xx-xx
+## [2.1.0] - xxxx-xx-xx
 
 ### Added
+- `no_overlap` argument in `sopa.aggregate` to avoid cells from overlapping when aggregating the channels
+- Add `unique_mapping` argument to map each bin to one unique cell for Visium HD data
 - Better documentation for `sopa.io.visium_hd` and a warning if the full res image is not loaded (#254)
 - `slideio` backend for WSI is now available and allows loading among others the Olympus `.vsi` format @stergioc
 - `nn.DataParallel` for WSI inference now allows using multiple GPUs and larger batch_sizes @stergioc
 
+
+### Changed
+- Use the `global` coordinate system by default in the remaining readers that were still using the `pixels` coordinate system
+- Use `prior_shapes_key: auto` in all Snakemake config - it will automatically find the right key based on the technology
+- Use `disk` from `skimage` for opening/closing in `sopa.segmentation.tissue`
+
+### Removed
+- Removed the `open-cv` dependency (#239)
+- Removed all deprecated functions that were announced to be removed in `sopa==2.1.0`
 
 ## [2.0.7] - 2025-05-19